fix: initialization sequence for `simpl-schema` et al
- Must load `Collection2` explicitly these days; see
https://atmospherejs.com/aldeed/collection2#import-using-static-imports
- Don't go the supposedly-new `/dynamic` route. Experimentation
suggests (although documentation fails to disclose) that
`Collection2.load()` is asynchronous for some reason; and we certainly
don't want Meteor to have a chance to start up before schemas are in
place (which it technically could, if we used a
`Collection2.load().then()` construct.) Again, experimentation
suggests (and documentation keeps mum on the fact) that using
`/static` imports, by the time `Meteor.startup` callbacks get to run,
`.attachSchema` happens to be available; just rely on that.
- Near as I can tell, there is no way to (synchronously) define and
share schemas using `unique:` with the npm package. One must use
`simpl-schema`-the-Atmosphere-package, which is
[bundled](https://atmospherejs.com/aldeed/collection2#40) inside
`aldeed:collection2` as of version 4.0.0 of the latter (and evidently
commingles with it to get “proper” startup ordering, as per above.)
I take it that
https://www.npmjs.com/package/simpl-schema#extending-the-schema-options
is what amounts to documentation for that fact.